Move SpecIcon
