+++ components/Home/SearchActiveFilters.vue
@@ -25,7 +25,7 @@
 <script>
 import { defineComponent } from '@nuxtjs/composition-api'
 import ActiveFilter from '~/components/molecules/ActiveFilter.vue'
-import SpecIcon from '~/components/Ui/SpecIcon.vue'
+import SpecIcon from '~/components/atoms/icons/SpecIcon.vue'
 import useGuildsStore from '~/composables/useGuildsStore'
 import specializationSlug from '~/data/utils/specializationSlug'
