+++ components/Home/SearchFilters.vue
@@ -44,7 +44,7 @@
 
 <script>
 import { defineComponent } from '@nuxtjs/composition-api'
-import SpecIcon from '~/components/Ui/SpecIcon.vue'
+import SpecIcon from '~/components/atoms/icons/SpecIcon.vue'
 import SelectCard from '~/components/atoms/SelectCard.vue'
 import useGuildsStore from '~/composables/useGuildsStore'
 import specializationSlug from '~/data/utils/specializationSlug'
