+++ pages/g/_slug.vue
@@ -103,7 +103,7 @@ import getClassName from '~/data/utils/getClassName'
 import getSpecName from '~/data/utils/getSpecName'
 import getClassTextColorClass from '~/data/utils/getClassTextColorClass'
 import BaseLoader from '~/components/atoms/BaseLoader.vue'
-import SpecIcon from '~/components/Ui/SpecIcon.vue'
+import SpecIcon from '~/components/atoms/icons/SpecIcon.vue'
 import ClockIcon from '~/components/atoms/icons/solid/ClockIcon.vue'
 import GlobeIcon from '~/components/atoms/icons/solid/GlobeIcon.vue'
 import CalendarIcon from '~/components/atoms/icons/solid/CalendarIcon.vue'
